+++ src/CSS/contact.css
@@ -1,0 +1,77 @@
+.container-contact{
+    width: 100%;
+    height: auto;
+    
+}
+
+.flex-header-contact{
+    display: flex;
+    justify-content: center;
+    font-size: 40px;
+    margin: 2%;
+}
+
+.flex-container-contact{
+    display: flex;
+    flex-direction: row;
+    width: 70%;
+    height: 100px;
+    margin-bottom: 80px;
+    margin-left: auto;
+    margin-right: auto;
+
+    transition: 0.5s;
+
+}
+
+.flex-items-contact{
+
+    flex-grow: 1;
+    text-align: center;
+    font-size: 50px;
+    transition: 0.5s;
+
+}
+
+.flex-items-contact a{
+    color: black;
+    text-decoration: none;
+}
+
+.flex-items-contact p{
+    font-size: 16px;
+}
+
+.flex-items-contact:hover{
+
+    transition: 0.5s;
+    transform: translateY(-10px);
+
+}
+
+@media screen and (max-width: 520px) {
+
+    .flex-container-contact{
+        display: flex;
+        flex-direction: column;
+        height: auto;
+        width: 100%;
+        transition: 0.5s;
+        justify-content: center;
+        align-items: center;
+    }
+    .flex-items-contact{
+        display: flex;
+        font-size: 40px;
+        min-width: 40%;
+        justify-content: space-between;
+        margin-left: 5%;
+    }
+
+
+    .flex-items-contact p {
+        display: inline;
+        padding-left: 50px;
+        padding-right: 0;
+    }
+}
