+++ src/CSS/nav.css
@@ -1,0 +1,86 @@
+body{
+  min-width: 200px;
+}
+.nav-container{
+    display: flex;
+    height: 90px;
+    width: 100%;
+    background-color: black;
+    color: white;
+    justify-content: space-between;
+    align-items: center;
+    /* position: fixed; */
+    top: 0;
+    z-index: 1;
+} 
+
+.active-list{
+    background-color: white;
+    height: 30px;
+    width: 90px;
+    color: black;
+    text-align: center;
+    border-radius: 2px;
+}
+
+.logo{
+    margin-left: 60px;
+    font-size: 20px;
+    cursor: pointer;
+}
+  
+.nav-menu{
+    display: flex;
+    padding-right: 60px;
+    margin-top: 1rem;
+}
+  
+.nav-menu li{
+    margin: 0 20px;
+    list-style-type: none;
+    font-size: 20px;
+    cursor: pointer;
+}
+
+
+.menu-bar{
+  font-size: 30px;
+  display: none;
+  padding-right: 60px;
+  padding-left: 60px;
+}
+
+@media screen and (max-width:780px) {
+
+  .nav-menu {
+    display: flex;
+    flex-direction: column;
+    position: absolute;
+    left: -100%;
+    top: 70px;
+    background-color: black;
+    width: 100%;
+    opacity: 0;
+    text-align: center;
+    padding: 0;
+    transition: 0.3s ease;
+    
+  }
+  .nav-menu li{
+    margin: 30px;
+  }
+
+  .nav-menu.active{
+    background-color: black;
+    left: 0;
+    opacity: 1;
+    z-index: 1;
+    transition: 0.3s ease;
+  }
+
+
+  .menu-bar{
+    display: flex;
+  }
+
+}
