+++ src/CSS/visual.css
@@ -1,0 +1,49 @@
+.size{
+    height: 200px;
+    width: auto;
+
+}
+
+h1{
+    font-weight: bold;
+}
+
+.visual-hover {
+    transition: 0.5s;
+
+}
+
+
+
+.view-visual{
+    display: flex;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+    justify-content: center;
+    align-items: center;
+    opacity: 0;
+    transition: 0.5s;
+}
+
+
+
+.visual-hover:hover {
+    cursor: pointer;
+    transform: translateY(-25px);
+    transition: 0.5s;
+}
+
+.visual-hover:hover img{
+    opacity: 0.7;
+}
+
+.visual-hover:hover .view-visual{
+
+    transition: 0.5s;
+    opacity: 1;
+
+}
