+++ src/CSS/work.css
@@ -1,0 +1,93 @@
+*{
+    box-sizing: border-box;
+}
+
+
+.container-work{
+    display: flex;
+    flex-direction: column;
+    height: 900px;
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+    background-color: white;
+}
+
+
+.top-container-work{
+    display: flex;
+    justify-content: center;
+    width: 100%;
+    height: auto;
+    gap: 50px;
+}
+
+.bottom-container-work{
+    display: flex;
+    width: 100%;
+    height: auto;
+    margin-top: 50px;
+    justify-content: center;
+    gap: 50px;
+}
+
+.flex-items-work{
+    display: flex;
+    flex-direction: column;
+    height: 300px;
+    width: 300px;
+    color: black;
+    border: 3px solid black;
+    border-radius: 5px;
+    justify-content: center;
+    text-align: center;
+    font-size: 70px;
+    transition: 0.5s;
+}
+
+.flex-items-work h1{
+    text-align: center;
+    font-size: 25px;
+}
+
+.container-header{
+    text-align: center;
+    font-size: 40px;
+    height: 100px;
+    width: 500px;
+    color: black;
+
+    
+}
+
+
+
+.flex-items-work:hover{
+    cursor: pointer;
+    transform: translateY(-25px);
+    transition: 0.5s;
+    color: white;
+    background-color: black;
+}
+
+@media screen and (max-width:800px){
+    .container-work{
+        height: 1200px;
+        
+    }
+    .top-container-work, .bottom-container-work{
+        justify-content: center;
+        align-items: center;
+        width: 60%;
+        flex-direction: column; 
+        /* border: 1px solid red; */
+    }
+
+    .flex-items-work{
+        height: 200px;
+        width: 200px;
+    }
+    
+}
+
+
