--- src/component/Footer.css
@@ -1,21 +1,0 @@
-footer{
-    width: 100%;
-    height: 50px;
-    color: white;
-    background-color: black;
-    text-align: center;
-    align-items: center;
-    margin: 0;
-}
-
-footer p{
-    transition: 0.5s;
-    padding-top: 15px;
-}
-
-@media screen and (max-width:500px){
-    p{
-        font-size: 10px;
-        transition: 0.5s;
-    }
-}
