--- src/component/Footer.js
@@ -1,9 +1,0 @@
-import "./Footer.css"
-
-function Footer(){
-    return(
-        <footer><p>Created by Chalermkiat Jaisuk | ©2023 All Rights Reserved</p></footer>
-    );
-}
-
-export default Footer;
