+++ src/component/Footer/index.js
@@ -1,0 +1,9 @@
+import "../../CSS/footer.css"
+
+function Footer(){
+    return(
+        <footer><p>Created by Chalermkiat Jaisuk | ©2023 All Rights Reserved</p></footer>
+    );
+}
+
+export default Footer;
