--- src/component/Home.css
@@ -1,46 +1,0 @@
-.container-home{
-    display: flex;
-    width: 100%;
-    height: 950px;
-    justify-content: center;
-    align-items: center;
-}
-
-
-
-.background{
-    display: flex;
-    position: relative;
-    background: url(https://images.unsplash.com/photo-1580927752452-89d86da3fa0a?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1170&q=80) no-repeat;
-    background-size: cover;
-    background-position: 0 50%;
-    width: auto;
-    height: auto;
-}
-
-.overlay{
-    width: 100%;
-    height: 100%;
-    background-color: rgb(0, 0, 0);
-    opacity: 0.5;
-    position: absolute;
-}
-
-.text-box{
-    position: absolute;
-    height: auto;
-    width: 50%;
-    text-align: center;
-    color: white;
-
-}
-
-.text-box h1{
-    margin-top: 0;
-    margin-bottom: 0;
-    font-size: 40px;
-}
-
-.text-box p{
-    font-size: 20px;
-}
