+++ src/index.js
@@ -2,24 +2,19 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "bootstrap/dist/css/bootstrap.min.css";
 import reportWebVitals from "./reportWebVitals";
-import Homepage from "./App";
-import "./index.css";
+import App from "./App";
+import "./CSS/index.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import Coding from "./component/Coding";
-import Production from "./component/Production";
-import Motion from "./component/motion";
-
-// import Home from "./component/Home";
-// import Contact from "./component/Contact";
-// import Work from "./component/Work";
-// import About from "./component/About";
-import Visual from "./component/Visual";
+import Coding from "./pages/coding/index";
+import Production from "./pages/production/index";
+import Motion from "./pages/motion/index";
+import Visual from "./pages/visual/index";
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <Homepage />,
+    element: <App />,
   },
 
   {
@@ -46,7 +41,4 @@ ReactDOM.createRoot(document.getElementById("root")).render(
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
